requirements.txt
openai>=1.0.0
requests>=2.28.0
python-dotenv>=0.20.0
rich>=12.0.0
pillow>=9.0.0
streamlit